fix: resolve all Pylance/Pyright lint errors across workspace
- Returning Any: wrap json.load, getattr, round/max/min, st.session_state
  returns with explicit casts, float(), str(), list(), or type: ignore
- Optional import fallbacks: type: ignore[assignment] for None sentinel
- terminal_spike_scanner: rename loop var q→quote to avoid type shadow
- tests/test_terminal: import ClassifiedItem at module level, annotate
  defaults dicts as dict[str, Any] for **-unpacking compatibility
- tests/test_benzinga_calendar: Generator return type for yield fixture
- streamlit_monitor: callable() check instead of truthiness for function

Files: 11 files, 0 errors remaining (excl. ~/.visidatarc external)
Tests: 1474 passed

+++ open_prep/run_open_prep.py
@@ -700,7 +700,7 @@ def _momentum_z_score_from_eod(candles: list[dict], period: int = 50) -> float:
         return 0.0
 
     z = (window[-1] - mean_ret) / std_ret
-    return round(max(min(z, 5.0), -5.0), 4)
+    return float(round(max(min(z, 5.0), -5.0), 4))
 
 
 def _enrich_quote_with_hvb(quote: dict[str, Any], hvb_multiplier: float = HVB_MULTIPLIER) -> None:
